src/test/java/TopCoder/SRM207_DIV2/CaptureThemAllTest.java
package TopCoder.SRM207_DIV2;

import static org.junit.Assert.assertEquals;

import java.io.FileInputStream;
import java.io.IOException;
import java.io.InputStream;
import java.lang.reflect.InvocationTargetException;
import java.lang.reflect.Method;
import java.lang.reflect.Modifier;
import java.nio.charset.StandardCharsets;
import java.util.Scanner;
import org.junit.Test;
import util.TestUtil;

public class CaptureThemAllTest{

  private final String TEST_NAME = "SRM207_DIV2";
  private final String PROBLEM_NAME = "1000";
  private final String METHOD_NAME = "fastKnight";
  
  @Test
  public void test1() throws IOException {doTest(1);}
  @Test
  public void test2() throws IOException {doTest(2);}
  @Test
  public void test3() throws IOException {doTest(3);}
  @Test
  public void test4() throws IOException {doTest(4);}
  @Test
  public void test5() throws IOException {doTest(5);}

  private void doTest(int testNo) throws IOException {

    InputStream fileInputStream = new FileInputStream (String.format("src\\test\\java\\TopCoder\\%s\\%s-%s-case%d.txt", TEST_NAME, TEST_NAME, PROBLEM_NAME, testNo));

    Scanner sc = new Scanner(fileInputStream);
    if(!sc.hasNext()){ sc.close(); return;}
    
    String knight = sc.nextLine().replace("\"", "");
    String rook = sc.nextLine().replace("\"", "");;
    String queen = sc.nextLine().replace("\"", "");;
    int answerVal = 0;

    try {
    	CaptureThemAll target = new CaptureThemAll();
      Class myClass = target.getClass();
      Method[] methods = myClass.getMethods();
      
      for(Method m: methods){
        if(Modifier.PUBLIC == m.getModifiers() && m.getName().equals(METHOD_NAME)){

          answerVal = ((Integer) m.invoke(target, knight, rook, queen)).intValue();

        }
      }
    } catch (IllegalAccessException | IllegalArgumentException | InvocationTargetException e) {
      e.printStackTrace();
    }finally{
      fileInputStream.close();
      sc.close();
    }

    int expetedVal = new Integer(TestUtil.dos2unix(TestUtil.readFile(String.format("src\\test\\java\\TopCoder\\%s\\%s-%s-case%d.answer.txt", TEST_NAME, TEST_NAME, PROBLEM_NAME, testNo),StandardCharsets.UTF_8).replace("\n", "").replace("\r", ""))).intValue();
    assertEquals(expetedVal, answerVal);

  };
}

